third_party/blink/web_tests/storage/indexeddb/legacy-constants-expected.txt



Verify that constants from previous version of the spec (beyond a grace period) have been removed:
PASS IDBKeyRange.SINGLE is undefined
PASS IDBKeyRange.LEFT_OPEN is undefined
PASS IDBKeyRange.RIGHT_OPEN is undefined
PASS IDBKeyRange.LEFT_BOUND is undefined
PASS IDBKeyRange.RIGHT_BOUND is undefined
PASS IDBTransaction.LOADING is undefined
PASS IDBRequest.LOADING is undefined
PASS IDBRequest.DONE is undefined
PASS IDBCursor.NEXT is undefined
PASS IDBCursor.NEXT_NO_DUPLICATE is undefined
PASS IDBCursor.PREV is undefined
PASS IDBCursor.PREV_NO_DUPLICATE is undefined
PASS IDBTransaction.READ_ONLY is undefined
PASS IDBTransaction.READ_WRITE is undefined
PASS IDBTransaction.VERSION_CHANGE is undefined
PASS successfullyParsed is true

TEST COMPLETE
